requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
charset-normalizer==3.4.6
Django==6.0.3
gunicorn==25.3.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
python-bidi==0.6.7
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
